+++ src/Designer/frontend/packages/schema-editor/src/components/SchemaInspector/ItemFieldsTab/ItemFieldsTable/ItemFieldsTableRow.tsx
@@ -2,13 +2,19 @@ import type { ReactNode, KeyboardEvent, ChangeEventHandler } from 'react';
 import { useState } from 'react';
 import classes from './ItemFieldsTable.module.css';
 import cn from 'classnames';
-import { deleteNode, setRequired, setPropertyName, type UiSchemaNode } from '@altinn/schema-model';
+import {
+  deleteNode,
+  duplicateNode,
+  setRequired,
+  setPropertyName,
+  type UiSchemaNode,
+} from '@altinn/schema-model';
 import { NameField } from '../../NameField';
 import { useSchemaEditorAppContext } from '@altinn/schema-editor/hooks/useSchemaEditorAppContext';
 import { AltinnConfirmDialog } from 'app-shared/components';
 import { useTranslation } from 'react-i18next';
-import { TrashIcon } from '@studio/icons';
-import { StudioCenter, StudioSwitch } from '@studio/components';
+import { FilesIcon, TrashIcon } from '@studio/icons';
+import { StudioButton, StudioCenter, StudioSwitch } from '@studio/components';
 import { nameFieldClass } from '@altinn/schema-editor/components/SchemaInspector/ItemFieldsTab/domUtils';
 import { ItemFieldType } from './ItemFieldType';
 
@@ -59,6 +65,10 @@ export const ItemFieldsTableRow = ({
     setSelectedUniquePointer(null);
   };
 
+  const duplicateHandler = () => {
+    save(duplicateNode(schemaModel, fullPath));
+  };
+
   return (
     <tr>
       <td className={cn(classes.tableColumnName, classes.tableCell)}>
@@ -84,6 +94,17 @@ export const ItemFieldsTableRow = ({
             disabled={readonly}
             name='checkedArray'
             onChange={changeRequiredHandler}
+          />
+        </StudioCenter>
+      </td>
+      <td className={cn(classes.tableColumnDuplicate, classes.tableCell)}>
+        <StudioCenter>
+          <StudioButton
+            title={t('schema_editor.duplicate_field')}
+            aria-label={t('schema_editor.duplicate_field')}
+            icon={<FilesIcon />}
+            onClick={duplicateHandler}
+            variant='tertiary'
           />
         </StudioCenter>
       </td>